Split resolver callable replay
